client/src/app/services/event/event.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { EventService } from './event.service';
import { IEventSchema, IPaginator, ISort, IEventResponse } from '../../interfaces';

describe('EventService', () => {

  let service: EventService;
  let httpMock: HttpTestingController;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [EventService]
    });
    service = TestBed.get(EventService);
    httpMock = TestBed.get(HttpTestingController);
  });

  afterEach(() => {
    httpMock.verify();
  });

  const dummyEvent = <IEventSchema> {
    _id: "1000",
    name: "Event 101"
  }

  const dummyPaginator = <IPaginator> {
    index: 1,
    size: 1,
    sort: <ISort> {field : "field", way : "way"}
  }
  
  const dummyEventResponse = <IEventResponse> {
    length: 1,
    list: [dummyEvent]
  }

  describe('getEvent()', () => {
    it('should return an Observable<IEventSchema>', () => {
      service.getEvent(dummyEvent._id).subscribe(event => {
        expect(event).toEqual(dummyEvent);
      });
  
      const req = httpMock.expectOne(`/${service.eventPath}/${dummyEvent._id}`);
      expect(req.request.method).toBe("GET");
      req.flush(dummyEvent);
    });
  });

  describe('createEvent()', () => {
    it('should return an Observable<IEventSchema>', () => {
      service.createEvent(dummyEvent).subscribe(event => {
        expect(event).toEqual(dummyEvent);
      });
  
      const req = httpMock.expectOne(`/${service.eventPath}`);
      expect(req.request.method).toBe("POST");
      req.flush(dummyEvent);
    });
  });

  describe('editEvent()', () => {
    it('should return an Observable<IEventSchema>', () => {
      service.editEvent(dummyEvent._id, dummyEvent).subscribe(event => {
        expect(event).toEqual(dummyEvent);
      });
  
      const req = httpMock.expectOne(`/${service.eventPath}/${dummyEvent._id}`);
      expect(req.request.method).toBe("PUT");
      req.flush(dummyEvent);
    });
  });

  // Don't know why this fails
  xdescribe('deleteEvent()', () => {
    it('should return an Observable<IEventResponse>', () => {
      service.deleteEvent(dummyEvent, dummyPaginator).subscribe(eventRes => {
        expect(eventRes).toEqual(dummyEventResponse);
      });
  
      const req = httpMock.expectOne({method: 'DELETE', url: `/${service.eventPath}/${dummyEvent._id}`});
      req.flush(dummyEventResponse);
    });
  });

});
